refactor(rc/controllers/ExerciseController.ts src/services/ExerciseService.ts src/routes/index.ts) excluimos o controller e service para adicionar no ja existente. mudamos a logica do findprogress e updatedprogress devido o schema esta diferente @mylenaverspeelt @stephanybrazeir0
